// src/schemas/blog.ts
import { z } from "zod";

export const blogSchema = z.object({
  title: z.string().min(3, "Title must be at least 3 characters long"),
  content: z.string().min(10, "Content must be at least 10 characters long"),
  image: z.instanceof(File),
});

export type TypeBlogSchema = z.infer<typeof blogSchema>;